% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_put_request.R
\name{bulk_put_request}
\alias{bulk_put_request}
\title{Utility to vectorize multiple requests}
\usage{
bulk_put_request(presigned_url, src)
}
\arguments{
\item{presigned_url}{A vector of URLs presigned by Document Cloud}

\item{src}{A vector of paths to documents to upload}
}
\value{
A content response
}
\description{
Utility to vectorize multiple requests
}
\examples{
bulk_put_request('http://documentcloud.org/document_is_here', './documents/document_to_upload.pdf' )
}
