% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filetype.R
\name{check_filetype}
\alias{check_filetype}
\title{Utility to test for valid filetype}
\usage{
check_filetype(path_list)
}
\arguments{
\item{path_list}{A list of paths to the document}
}
\value{
A list of valid extensions according to filetype
}
\description{
Utility to test for valid filetype
}
\examples{
list_of_paths <- c('not_a_pdf.zip', 'is_a_pdf.pdf', 'is_a_jpeg.jpg')
check_filetype(list_of_paths)
}
