% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_refresh_token.R
\name{dc_refresh_token}
\alias{dc_refresh_token}
\title{Refresh authentication tokens}
\usage{
dc_refresh_token(refresh_token)
}
\arguments{
\item{refresh_token}{Content response of refresh toekn request}
}
\value{
Response of content request
}
\description{
Refresh authentication tokens
}
\examples{
refresh_token <- 'hErE_Is_A_rEfReSh_ToKeN'
refresh_auth <- dc_refresh_token(refresh_token)
}
