% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_documents.R
\name{upload_documents}
\alias{upload_documents}
\title{Upload files to DocumentCloud}
\usage{
upload_documents(path_list, project_id, refresh_token)
}
\arguments{
\item{path_list}{A list of paths to the documents}

\item{project_id}{A project ID from DocumentCloud}

\item{refresh_token}{Authentication refresh token from DocumentCloud}
}
\value{
A dataframe of paths and destination URLs
}
\description{
Upload files to DocumentCloud
}
\examples{
refresh_token <- 'hErE_Is_A_rEfReSh_ToKeN'
file_names <- c('not_a_pdf.zip', 'is_a_pdf.pdf', 'is_a_jpeg.jpg')
upload_documents(file_names, 208620, refresh_token)
}
